Sequence - Update dependencies
Summary:
This is a remake of D13 as it broke.

This fixes T91.

build.gradle is update to use the newest version of libaries
A call to ".getImage" on a pictogram model have been replaced with using the controller, as the old api is removed.

Test Plan:
Verify that the sequence app works (This is only the parts of the app excluding the view used by
citizens)

Reviewers: jejens13

Reviewed By: jejens13

Maniphest Tasks: T91

Differential Revision: http://web.giraf.cs.aau.dk/D41